backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pandas==2.2.1
scikit-learn==1.4.1.post1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.3
numpy==1.26.4
